requirements.txt
Django==1.10.4
django-bootstrap3==7.1.0
guessit==2.1.1
isodate==0.5.4
patool==1.12
pytz==2016.7
trakt.py==2.9.0
